Move the mocks to their own separate subdirectory / submodule.
