+++ local-modules/doc-common/mocks/index.js
@@ -1,0 +1,13 @@
+// Copyright 2016-2017 the Bayou Authors (Dan Bornstein et alia).
+// Licensed AS IS and WITHOUT WARRANTY under the Apache License,
+// Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
+
+import MockChange from './MockChange';
+import MockDelta from './MockDelta';
+import MockOp from './MockOp';
+
+export {
+  MockChange,
+  MockDelta,
+  MockOp
+};
